LayoutTests/platform/ios-sim-deprecated/fast/events/spatial-navigation/snav-offscreen-content-expected.txt

offscreen contentRight link

Down link
FAIL document.activeElement.getAttribute("id") should be 1. Was start.
